requirements.txt
numpy==1.25.2
opencv-python
opencv-python-headless
onnxruntime-gpu
torch
torchvision